Fix plt.show() and refactor launch into main
Created a new main file to hold the lauching process before the main
window is created. Also fixes the issue with plt.show() not working by
setting the global figure manager as early as possible.

RE #27374
