+++ qt/applications/workbench/setup.py.in
@@ -16,7 +16,7 @@ setup(
     package_data={'': ['*.ui']},
     entry_points={
         'gui_scripts': [
-            'workbench = workbench.app.mainwindow:main'
+            'workbench = workbench.app.main:main'
         ]
     },
     @SETUPTOOLS_BUILD_COMMANDS_USE@
