+++ qt/applications/workbench/workbench/plotting/config.py
@@ -33,12 +33,17 @@ def initialize_matplotlib():
     :param figure_window_parent: An QWidget that will become the parent of any figure window. Can be None
     :param figure_window_flags: A Qt.WindowFlags enumeration defining the window flags for a figure window
     """
-    # Replace vanilla Gcf with our custom manager
-    setattr(_pylab_helpers, 'Gcf', GlobalFigureManager)
     # Set our defaults
     reset_rcparams_to_default()
     # Set figure DPI scaling to monitor DPI
     mpl.rcParams['figure.dpi'] = QApplication.instance().desktop().physicalDpiX()
+
+
+def init_mpl_gcf():
+    """
+    Replace vanilla Gcf with our custom manager
+    """
+    setattr(_pylab_helpers, 'Gcf', GlobalFigureManager)
 
 
 def reset_rcparams_to_default():
